+++ src/mastodon_sim/engines/__init__.py
@@ -1,7 +1,7 @@
 """Engine extension package."""
 
 from mastodon_sim.engines.base import BaseEnvironmentEngine
-from mastodon_sim.engines.runtime import (
+from mastodon_sim.engines.base_engines import (
     BaseRuntimeEngine,
     FlowRuntimeEngine,
 )
