+++ src/mastodon_sim/engines/multi_gm.py
@@ -18,7 +18,7 @@ from typing import Any
 from concordia.typing import entity as entity_lib
 from typing_extensions import override
 
-from mastodon_sim.engines.runtime import FlowRuntimeEngine
+from mastodon_sim.engines.base_engines import FlowRuntimeEngine
 from mastodon_sim.runtime.config import ConfigStore
 
 _LOGGER = logging.getLogger(__name__)
