+++ src/mastodon_sim/runtime/factories.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from omegaconf import DictConfig
 
-from mastodon_sim.engines.runtime import (
+from mastodon_sim.engines.base_engines import (
     BaseRuntimeEngine,
     FlowRuntimeEngine,
 )
